Solaris.
svn:sync-xref-src-repo-rev: r24073

+++ src/VBox/Runtime/r0drv/solaris/time-r0drv-solaris.c
@@ -1,4 +1,4 @@
-/* $Id: time-r0drv-solaris.c 4287 2007-08-22 14:49:19Z knut.osmundsen@oracle.com $ */
+/* $Id: time-r0drv-solaris.c 4474 2007-08-31 19:23:49Z knut.osmundsen@oracle.com $ */
 /** @file
  * innotek Portable Runtime - Time, Ring-0 Driver, Solaris.
  */
@@ -27,7 +27,7 @@
 
 RTDECL(uint64_t) RTTimeNanoTS(void)
 {
-    return gethrtime();
+    return (uint64_t)gethrtime();
 }
